rename 'queue_on_list' to 'queue_by_list' to match the class name
